refactor(ui): reorganize DatabaseListModules sub-components into folders
Move sub-components to follow Component Folder Structure guidelines:
- Create dedicated folder for each sub-component
- Extract interfaces to .types.ts files
- Add components/index.ts barrel file

References: #RI-7781
