+++ redisinsight/ui/src/components/database-list-modules/components/DatabaseModuleContent/DatabaseModuleContent.types.ts
@@ -1,0 +1,5 @@
+import type { IDatabaseModule } from 'uiSrc/utils'
+
+export interface DatabaseModuleContentProps {
+  modules: IDatabaseModule[]
+}
