+++ redisinsight/ui/src/components/database-list-modules/components/DatabaseModuleContentItem/DatabaseModuleContentItem.types.ts
@@ -1,0 +1,8 @@
+import type { AllIconsType } from "uiSrc/components/base/icons"
+
+export interface DatabaseModuleContentItemProps {
+  icon?: AllIconsType
+  content?: string
+  abbreviation?: string
+}
+
