+++ redisinsight/ui/src/components/database-list-modules/components/DatabaseModuleItem/DatabaseModuleItem.types.ts
@@ -1,0 +1,10 @@
+import type { AllIconsType } from "uiSrc/components/base/icons"
+
+export interface DatabaseModuleItemProps {
+  abbreviation?: string
+  icon?: AllIconsType | null
+  content?: string
+  inCircle?: boolean
+  onCopy?: (text: string) => void
+}
+
