+++ redisinsight/ui/src/components/database-list-modules/components/DatabaseModulesList/DatabaseModulesList.types.ts
@@ -1,0 +1,8 @@
+import type { IDatabaseModule } from 'uiSrc/utils/modules'
+
+export interface DatabaseModulesListProps {
+  modules: IDatabaseModule[]
+  contentItems: IDatabaseModule[]
+  inCircle?: boolean
+  anchorClassName?: string
+}
