+++ redisinsight/ui/src/components/database-list-modules/components/index.ts
@@ -1,0 +1,11 @@
+export { DatabaseModuleContent } from './DatabaseModuleContent/DatabaseModuleContent'
+export type { DatabaseModuleContentProps } from './DatabaseModuleContent/DatabaseModuleContent.types'
+
+export { DatabaseModuleContentItem } from './DatabaseModuleContentItem/DatabaseModuleContentItem'
+export type { DatabaseModuleContentItemProps } from './DatabaseModuleContentItem/DatabaseModuleContentItem.types'
+
+export { DatabaseModuleItem } from './DatabaseModuleItem/DatabaseModuleItem'
+export type { DatabaseModuleItemProps } from './DatabaseModuleItem/DatabaseModuleItem.types'
+
+export { DatabaseModulesList } from './DatabaseModulesList/DatabaseModulesList'
+export type { DatabaseModulesListProps } from './DatabaseModulesList/DatabaseModulesList.types'
